LayoutTests/storage/indexeddb/basics.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test IndexedDB's basics.");
if (window.layoutTestController)
    layoutTestController.waitUntilDone();

function test()
{
    shouldBeTrue("'webkitIndexedDB' in window");
    shouldBeFalse("webkitIndexedDB == null");

    // FIXME: Verify other IndexedDatabaseRequest constructors, once they're implemented.

    result = evalAndLog("webkitIndexedDB.open('name')");
    verifyResult(result);
    result.onsuccess = openCallback;
    result.onerror = unexpectedErrorCallback;
}

function openCallback()
{
    verifySuccessEvent(event);
    done();
}


test();

var successfullyParsed = true;

</script>
</body>
</html>
